Перешли на использование CrudRepository.java

+++ src/main/java/library/model/User.java
@@ -5,19 +5,19 @@ package library.model;
 
 
 public class User {
-    private Integer id;
+    private Long id;
     private String name;
 
-    public User(final Integer id, final String name) {
+    public User(final Long id, final String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(final Integer id) {
+    public void setId(final Long id) {
         this.id = id;
     }
